error checking and bug fixes

+++ src/inventoryManagementSystem/DbInitializer.java
@@ -30,7 +30,7 @@ public class DbInitializer {
 			statement.execute("""
 					CREATE TABLE IF NOT EXISTS product (
 						id INT PRIMARY KEY AUTO_INCREMENT,
-						price INT NOT NULL,
+						price FLOAT NOT NULL,
 						quantity INT NOT NULL,
 						name VARCHAR(255) NOT NULL,
 						description VARCHAR(255)
